Add files via upload
Scientific Question: In terms of presenting greater face and construct validities to study human hereditary hemochromatosis, do mice (Mus musculus) models have a greater sequence similarity (than rat (Rattus norvegicus) models) with the human (Homo sapiens) variants of the HFE, Tfr1, and Tfr2 proteins?

Hypothesis: If the Mus musculus variants of the HFE, Tfr1, and Tfr2 genes share a greater sequence similarity with the human variants, mice models are a better choice (in comparison to Rattus norvegicus models) to study human hereditary hemochromatosis.

The HFE, Tfr1, and Tfr2 proteins are some of the most important ones involved in the causation of hereditary hemochromatosis in humans. HFE is a homeostatic iron regulator protein. Tfr1 and Tfr2 are two different variants of transferrin receptor proteins. Mutations in one or more of these proteins are commonly considered to be responsible for the hemochromatosis phenotype. In this context, I aimed to study the similarities of these proteins in their natural forms between human and mice/rats in order to explore if any one of the animal model offers a greater similarity over the other. This is done under the assumption that having a model with protein variants that have greater similarity to the human versions would offer an overall better construct and face validities to the experiment/study.